I finished

+++ dmitriy/src/main/hw/hw2/TaskFive.java
@@ -12,7 +12,7 @@ public class TaskFive {
     public static void main(String[] args) {
         float f = (float)128.50;
         int i = (int)f;
-        int b = (int)(i + f);
+        int b = (byte)(i + f);
         System.out.println(b);
     }
 }
